+++ hypatia/constants.py
@@ -1,17 +1,28 @@
 # This module is part of Hypatia and is released under the
 # MIT License: http://opensource.org/licenses/MIT
 
-"""Constants for Hypatia Engine
+"""Constants, enumerations for Hypatia.
 
-Right now this just contains stuff pertaining to Walkabout, i.e.,
-direction and action specifications. These classes are to be passed as
-parameters.
+Generally, these constants and enumerations serve to replace
+the usage of strings for such paramount attributes like
+direction and action. The benefit of this model is:
 
-Examples:
+  * enhances code clarity
+  * provides type checking
+  * provides constants operations, e.g.,
+    `North + East = North East`.
+  * class methods to convert one datatype to a constant, like
+    velocity to direction
+
+Example:
   >>> from hypatia import animations
   >>> sprite = animations.Walkabout('debug')
   >>> sprite.animations[Action.walk][Direction.east]
   <pyganim.PygAnimation object at 0x...>
+
+See Also:
+   *  :attribute:`actor.Actor.direction`
+   *  :attribute:`animations.Walkabout.direction`
 
 """
 
@@ -24,8 +35,8 @@ import enum
 class Direction(enum.Enum):
     """Compass direction. Specific to movement of a sprite/surface.
 
-    Inspired by Unix numerical permissions. Only ever combined with
-    one other direction max.
+    Inspired by Unix numerical permissions. Only ever
+    combined with one other direction max.
 
     See Also:
         :class:`physics.Velocity`
@@ -63,33 +74,68 @@ class Direction(enum.Enum):
 
     @classmethod
     def cardinal(cls):
+        """Return a tuple of the cardinal directions in the order:
+        North, East, South, West.
+
+        Returns:
+            tuple: (north, east, south, west)
+
+        """
 
         return (cls.north, cls.east, cls.south, cls.west)
 
     @classmethod
     def x_plus(cls):
+        """Returns the direction associated
+        with moving RIGHT (+x) on the X-AXIS.
+
+        Returns:
+            Direction.east
+
+        """
 
         return cls.east
 
     @classmethod
     def x_minus(cls):
+        """Returns the direction associated
+        with moving LEFT (-x) on the X-AXIS.
+
+        Returns:
+            Direction.west
+
+        """
 
         return cls.west
 
     @classmethod
     def y_plus(cls):
+        """Returns the direction associated
+        with moving DOWN (+y) on the Y-AXIS.
+
+        Returns:
+            Direction.south
+
+        """
 
         return cls.south
 
     @classmethod
     def y_minus(cls):
+        """Returns the direction associated
+        with moving UP (-y) on the Y-AXIS.
+
+        Returns:
+            Direction.north
+
+        """
 
         return cls.north
 
     @classmethod
     def from_velocity(cls, velocity):
-        """Return a direction which corresponds to the current 2D
-        velocity.
+        """Return a direction which corresponds
+        to the current 2D velocity.
 
         See Also:
             :class:`constants.Direction`
@@ -99,28 +145,50 @@ class Direction(enum.Enum):
 
         """
 
+        # We're going to combine the directions
+        # extrapolated from each axis, then
+        # combine them to make a new direction!
         collected_directions = []
 
         for axis in ['x', 'y']:
+            # e.g., call Direction.x_plus() to get the positive
+            # axis direction for 'x' (which would be East).
             plus_direction = getattr(Direction, axis + '_plus')()
+
+            # e.g., call Direction.y_minus() to get the negative
+            # axis direction for 'y' (which would be North).
             minus_direction = getattr(Direction, axis + '_minus')()
+
+            # get the current velocity for this axis, determine
+            # if it's positive (use plus_direction), negative
+            # (use minus_direction) or neutral (do nothing!).
             axis_value = getattr(velocity, axis)
 
             if axis_value > 0:
+                # the velocity for this axis is positive,
+                # therefore it is the "plus direction."
                 collected_directions.append(plus_direction)
             elif axis_value == 0:
+                # the velocity for this axis is neutral,
+                # therefore we cannot extrapolate
+                # direction from velocity.
                 pass
             else:
+                # Deductively, the axis value is negative,
+                # therefore it is the "minus_direction."
                 collected_directions.append(minus_direction)
 
+        # Cool trick, huh? North + East = North East, so forth.
+        # Be sure to check out Direction.__add__.
         return collected_directions[0] + collected_directions[1]
 
     def __add__(cls, other_direction):
-        """Combine one cardinal direction with another to get
-        an ordinal direction.
+        """Combine one cardinal direction with
+        another to get an ordinal direction.
 
         Args:
-            other_direction (:class:`Direction`): --
+            other_direction (Direction): one of the Direction
+                enumerations.
 
         Returns:
             :class:`Direction`: an ordinal direction.
@@ -137,6 +205,14 @@ class Direction(enum.Enum):
 @enum.unique
 class Action(enum.Enum):
     """Specific to movement of a sprite/surface.
+
+    Attributes:
+        stand (int): Actor standing/normal/no-input state.
+        walk (int): Actor walking/moving state. Actor
+            has velocity.
+
+    See Also:
+        :class:`animations.Walkabout`
 
     """