+++ hypatia/actor.py
@@ -3,20 +3,28 @@
 
 """Implementation of actors.
 
-Actors are the base of any entities which may perform actions. Examples
-of actors are enemies, NPCs, the player themselves, etc. This module
-implements a basic :class:`Actor` class to serve as the parent for any
-classes representing objects such those examples. Any logic which can
-be shared by all objects belongs in the class, for example, the logic
-for moving around the game world. This approach makes it possible to
-allow, for example, enemimes and players to share much of the same
-behavior, a nd this can be useful by making it easer (from a
-programming point-of-view) to give monsters the same set of core
-abilities and logic as NPCs, etc.
+Actors are the base of any entities (players) which may perform
+actions, examples include:
+
+  * human player
+  * an enemy NPC
+  * a friendly NPC
+  * an invisible NPC which simply displays
+    a message when "talked" to.
+
+This module implements a basic :class:`Actor` class to serve as
+the parent for any classes representing objects such as those
+examples.
+
+Any logic which can by shared by all "players" belongs in the class,
+for example, the logic for moving around the game world. This approach
+makes it possible to allow, for example, enemies and players to share
+much of the same behavior and this can make it easier to give monsters
+the same set of core abilities and logic as NPCs, etc.
 
 When type-checking is necessary the :class:`Actor` class provides a
 useful way to test for objects which support a bare-minimum of core,
-shared actions.  The class is also useful in role-playing games for
+shared actions. The class is also useful in role-playing games for
 storing data that tends to be common between the Player, NPCs, enemies,
 etalia, a common example being statistics like hit-points.
 
@@ -30,38 +38,48 @@ from hypatia import physics
 class Actor(object):
     """The base class for any entity which can perform actions.
 
-    For example, both :class:`player.Player` and :class:`player.NPC`
-    objects can move around the game world. This is the type of action
-    which is shared by all "actors" and therefore best implemented in
-    this class, allowing it to be shared by as many entities as
-    possible, e.g. enemies.
+    For example, most actors can move around the game world, so
+    there are tools for setting the "direction" the actor is
+    facing. Another example: most actors can say something, offer
+    some dialog. These are the types of actions which are shared
+    by all/most actors and therefore best implemented in this class,
+    allowing it to be shared by as many players as possible, e.g.,
+    human player, enemies.
 
-    It is typically not useful to directly instantiate :class:`Actor`
-    objects but the implementation does not prevent this.
+    Note:
+      It is typically not useful to directly instantiate :class:`Actor`
+      objects but the implementation does not prevent this.
 
     Attributes:
-        walkabout (:class:`animations.Walkabout`): instance.
+        walkabout (animations.Walkabout): --
+        direction (constants.Direction): --
+
+    See Also:
+        :mod:`actor`
 
     """
 
-    def __init__(self, walkabout, say_text=None, velocity=None):
+    def __init__(self, walkabout=None, say_text=None, velocity=None):
         """Constructs a new Actor.
 
         Args:
-            walkabout (:class:`animations.Walkabout`): Walkabout which
-                is then accessible via the ``walkabout`` property. This
-                argument is optional and defaults to new instance
-                of :class:`animations.Walkabout`.
+            walkabout (Optional[animations.Walkabout]): Optionally
+                set a walkabout property, which will graphically
+                represent the actor.
+            say_text (Optional[str]): Optionally set the text which
+                is displayed when this actor's :meth:`Actor.say()`
+                is called.
+            velocity (Optional[physics.Velocity]): --
 
         """
 
         self.walkabout = walkabout
-        self.say_text = say_text or None
+        self.say_text = say_text
         self.velocity = velocity or physics.Velocity()
 
     @property
     def direction(self):
-        """An intsance of :class:`constants.Direction`
+        """An instance of :class:`constants.Direction`
 
         This property indicates the direction the actor is facing.
         Is it possible to set this property to a new value.
@@ -72,10 +90,22 @@ class Actor(object):
             TypeError: If one tries to delete this property
 
         """
+
         return self.walkabout.direction
 
     @direction.setter
     def direction(self, new_direction):
+        """Set the direction this actor is facing.
+
+        Args:
+            new_direction (constants.Direction): The new direction
+                for the actor to face.
+
+        Raises:
+            AttributeError: If the new value is not a valid object
+                of the :class:`constants.Direction` class.
+
+        """
 
         if not isinstance(new_direction, constants.Direction):
 
@@ -87,10 +117,39 @@ class Actor(object):
 
     @direction.deleter
     def direction(self):
+        """You are not allowed to delete the direction of an Actor.
+
+        Raises:
+            TypeError: If one tries to delete this property
+
+        """
 
         raise TypeError("Cannot delete the 'direction' of an Actor")
 
     def say(self, at_direction, dialogbox):
+        """Change this actor's direction, and say this actor's
+        say_text in the global dialog box.
+
+        This method is typically called by another
+        actor's :meth:`actor.Actor.talk()`.
+
+        Args:
+            at_direction (constants.Direction): The new direction
+                for this actor to face.
+            dialogbox (dialog.DialogBox): the DialogBox to print
+                this actor's say_text to.
+
+        Returns:
+            bool: True if this actor can and did say something, False
+                if this actor cannot say something. This quality is
+                determined by the presence of say_text.
+
+        Note:
+            Even if this actor doesn't say anything, it will
+            change the direction it's facing.
+
+        """
+
         facing = {
                   constants.Direction.north: constants.Direction.south,
                   constants.Direction.east: constants.Direction.west,
@@ -102,8 +161,28 @@ class Actor(object):
         if self.say_text:
             dialogbox.set_message(self.say_text)
 
+            return True
+
+        else:
+
+            return False
+
     def talk(self, npcs, dialogbox):
-        """Attempt to talk in current direction.
+        """Trigger another actor's :meth:`actor.Actor.say()` if
+        they are immediately *in front* of this actor.
+
+        See Also:
+            * :attribute:`animations.Walkabout.direction`
+            * :attribute:`Actor.direction`
+            * :meth:`actor.Actor.say()`
+
+        Args:
+            npcs (List[player.Npc]): NPCs to check for
+                collisions immediately in front of this
+                actor.
+            dialogbox (dialog.DialogBox): The dialogbox which
+                another actor will print to if they have
+                something to say.
 
         """