+++ hypatia/tiles.py
@@ -8,7 +8,8 @@ which consists of graphical tiles aligned to a grid. Provides tools for
 loading specific tile resources into an object. Contains information
 about tiles (tile properties).
 
-For more information see: http://en.wikipedia.org/wiki/Tile_engine
+See Also:
+    http://en.wikipedia.org/wiki/Tile_engine
 
 """
 
@@ -27,14 +28,16 @@ from hypatia import animations
 
 
 class BadTileID(Exception):
-    """Tilesheet: tile was referenced by an ID which does not exist.
+    """Tilesheet: tile was referenced by an
+    ID which does not exist.
 
     Args:
-      bad_tile_id (int): the tile id referenced which does not actually
-        exist in a Tilesheet.
+        bad_tile_id (int): the tile id referenced which
+            does not actually exist in a Tilesheet.
 
     Attributes:
-      bad_tile_id (int): the tile ID referenced which does not exist.
+        bad_tile_id (int): the tile ID referenced
+            which does not exist.
 
     """
 
@@ -146,6 +149,10 @@ class TileMap(object):
         self.impassable_rects = impassable_rects
         self.animated_tile_stack = animated_tile_stack
         self.dimensions_in_tiles = dimensions_in_tiles
+
+        # the 3D list of Tilesheet tile IDs
+        # which constructed this TileMap. It
+        # is not updated when self.tiles is.
         self._tile_ids = tile_ids
 
     def __getitem__(self, coord):
@@ -198,6 +205,15 @@ class TileMap(object):
         return self[(tile_x, tile_y)]
 
     def blit_layer_animated_tiles(self, viewport, layer):
+        """Blit all of the animated tiles from a
+        designated layer to the supplied viewport.
+
+        Args:
+            viewport (render.Viewport): --
+            layer (int): The nth layer of animated tiles
+                which to blit to viewport.
+
+        """
 
         for tile_pyganim, position in self.animated_tile_stack[layer]:
             tile_pyganim.blit(viewport.surface,
@@ -225,8 +241,13 @@ class TileMap(object):
     def to_string(self, separator=' '):
         """Create the user-unfriendly string for the tilemap.
 
+        Used for creating tilemap.txt.
+
         Args:
           separator (str): can be ''
+
+        Returns:
+            str: --
 
         """
 
@@ -257,6 +278,11 @@ class TileMap(object):
     def from_string(cls, map_string, separator=' '):
         """This is a debug feature. Create a 3D list of tile names using
         ASCII symbols. Supports layers.
+
+        Used for reading tilemap.txt.
+
+        Returns:
+            TileMap: --
 
         """